src/requirement.txt
Django==3.2.9
djangorestframework==3.12.4
django-rest-knox==4.2.0
channels==3.0.4
channels-redis==3.3.1
psycopg2-binary==2.9.1
daphne==3.0.2
asgiref==3.3.4
